+++ src/backend/marsha/core/utils/send_emails.py
@@ -41,3 +41,28 @@ def send_ready_to_convert_notification(video):
         )
 
     return None
+
+
+def send_vod_ready_notification(video):
+    """Send a convert to vod ready email notification."""
+    if email := video.live_stopped_by_email:
+        template_vars = {
+            "email": email,
+            "time_zone": settings.TIME_ZONE,
+            "video": video,
+            "video_access_url": video.last_lti_url,
+        }
+
+        msg_plain = render_to_string("core/mail/text/vod_ready.txt", template_vars)
+        msg_html = render_to_string("core/mail/html/vod_ready.html", template_vars)
+        return send_mail(
+            subject=_("Live Stream is converted to VOD - %(video.title)s")
+            % {"video.title": video.title},
+            message=msg_plain,
+            from_email=None,
+            recipient_list=[email],
+            html_message=msg_html,
+            fail_silently=False,
+        )
+
+    return None