+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Send an email when a live stream is ready to be converted to VOD
+- Send an email when a live stream is converted to VOD
 
 ### Changed